Fix errors with custom parameters
The way custom parameters are initiated is changed in a recent update in
zhinst-toolkit. When an AWG module is initiated in zhinst-toolkit, the
custom parameters are initiated as `None` in the beginning. After the
connection is established with the device, the parameter information is
extracted directly from the device node, with the `_init_awg_params`
method. Therefore, zhinst-qcodes also needs to be updated to call this
method explicitly. After calling this method, the custom parameters can
be added as QCoDeS parameters.